refactor: skills placed in Skills component
