+++ frontend/src/pages/identity/UserProfile.tsx
@@ -10,11 +10,10 @@ import {
   FormErrorMessage,
   FormControl,
   InputGroup,
-  Select,
 } from '@chakra-ui/react';
 
 import {
-  FC, useCallback, useEffect, useRef, useState,
+  FC, useCallback, useEffect, useRef,
 } from 'react';
 import { useUser } from 'hooks/useUser';
 import {
@@ -31,7 +30,7 @@ import useCountries from 'hooks/useCountries';
 import useTitleOptions from 'hooks/useTitleOptions';
 import useTimezoneOptions from 'hooks/useTimezoneOptions';
 import useBusinessAreaOptions from 'hooks/useBusinessAreaOptions';
-import useSkillsOptions from 'hooks/useSkillsOptions';
+import Skills from 'components/user-profile-components/Skills';
 
 const UserProfile: FC = () => {
   const { user } = useUser();
@@ -39,9 +38,6 @@ const UserProfile: FC = () => {
   const titleOptions = useTitleOptions();
   const timezoneOptions = useTimezoneOptions();
   const businessAreaOptions = useBusinessAreaOptions();
-  const skillsOptions = useSkillsOptions()!;
-  const [userSkillList, setSkillList] = useState(user?.profile.skills!);
-  const [addedArea, setAddedArea] = useState('-');
 
   const { t } = useTranslation();
 
@@ -75,19 +71,6 @@ const UserProfile: FC = () => {
   }, [userProfile, reset]);
 
   const avatarFile = watch('avatar');
-
-  const deleteSkillHandler = (item:any) => {
-    const filtered = userSkillList.filter((value: any) => value !== item);
-    setSkillList(filtered);
-  };
-
-  const addSkillHandler = () => {
-    if (userSkillList.includes(addedArea) || addedArea === '') {
-      return;
-    }
-    const newList = userSkillList.concat([addedArea]);
-    setSkillList(newList);
-  };
 
   return (
     <form
@@ -197,32 +180,7 @@ const UserProfile: FC = () => {
             type="number"
           />
 
-          <FormLabel>Current skills</FormLabel>
-          <Stack spacing={2}>
-            {userSkillList
-              .map((item) => (
-                <Stack direction="row" spacing={4}>
-                  <li key={item} value={item}>{item}</li>
-                  <Button onClick={() => deleteSkillHandler(item)} size="sm" colorScheme="red">Delete</Button>
-                </Stack>
-              ))}
-          </Stack>
-          <FormLabel>Add skills from list below</FormLabel>
-
-          <Stack direction="row">
-            <Select
-              onChange={(e) => {
-                const val = e.target.value;
-                setAddedArea(val);
-              }}
-              value={addedArea}
-            >
-              {skillsOptions.map((item) => (
-                <option key={item.value} value={item.label}>{item.value}</option>
-              ))}
-            </Select>
-            <Button onClick={() => addSkillHandler()} size="sm" colorScheme="green">Add</Button>
-          </Stack>
+          <Skills />
 
           <FormControl id="non-field" isInvalid={Boolean(errors?.non_field_errors)} mt={['0 !important']}>
             <FormErrorMessage>{errors?.non_field_errors}</FormErrorMessage>
